Complete front end and finish tests.

+++ db/schema/billingSchema.js
@@ -6,16 +6,10 @@ const { seeds } = require('../sql/seed');
 
 const BillingRate = db.define('billing_rate', {
   country: Sequelize.STRING,
-  basic_rate: Sequelize.INTEGER,
-  expedited_rate: Sequelize.INTEGER,
-  one_day_rate: Sequelize.INTEGER
+  basic_rate: Sequelize.DECIMAL(10, 2),
+  expedited_rate: Sequelize.DECIMAL(10, 2),
+  one_day_rate: Sequelize.DECIMAL(10, 2)
 });
-
-// const ZipCode = db.define('zip_code', {
-//   zip: Sequelize.INTEGER,
-//   city: Sequelize.STRING,
-//   state: Sequelize.STRING
-// });
 
 BillingRate.sync({force: true})
   .then(() => {
@@ -25,16 +19,6 @@ BillingRate.sync({force: true})
   .catch(err => {
     console.log('error connecting to database ', err);
   });
-
-// ZipCode.sync({force: true})
-//   .then(() => {
-//     ZipCode.bulkCreate(zips);
-//     console.log('table created for zipcode');
-//   })
-//   .catch(err => {
-//     console.log('error connecting to database');
-//   });
-
 
 module.exports = {
   BillingRate: BillingRate
